3.9: fixed merging
